chore: Move movesRedirect method from home controller to middleware
