+++ app/home/controllers.js
@@ -1,19 +1,4 @@
-const { get } = require('lodash')
-
 const dateHelpers = require('../../common/helpers/date')
-const permissions = require('../../common/middleware/permissions')
-
-function movesRedirect(req, res, next) {
-  // TODO: Remove this once we enable the dashboard for all users
-  // Moves will then likely always redirect to the dashboard
-  const userPermissions = get(req.session, 'user.permissions')
-
-  if (!permissions.check('dashboard:view', userPermissions)) {
-    return res.redirect('/moves')
-  }
-
-  next()
-}
 
 function dashboard(req, res) {
   const currentWeek = dateHelpers.getCurrentWeekAsRange()
@@ -55,6 +40,5 @@ function dashboard(req, res) {
 }
 
 module.exports = {
-  movesRedirect,
   dashboard,
 }
